Heater System code
In this code the behavior of Model Predictive Control can be examined. Based on the control Horizon and Plant Model the optimizer gives out a set of in put controls which will give out corresponding Outputs. Note that this code only shows the behavior of the system in just one time step.